+++ scripts/test/GenericInterface/Operation/Ticket/TicketSearch.t
@@ -7,6 +7,7 @@
 # did not receive this file, see http://www.gnu.org/licenses/agpl.txt.
 # --
 
+## no critic (Modules::RequireExplicitPackage)
 use strict;
 use warnings;
 use utf8;
@@ -297,6 +298,24 @@ my $TicketID1 = $TicketObject->TicketCreate(
 $Self->True(
     $TicketID1,
     "TicketCreate() successful for Ticket One ID $TicketID1",
+);
+
+# update escalation times directly in the DB
+my $EscalationTime = $TimeObject->SystemTime() + 120;
+return if !$Self->{DBObject}->Do(
+    SQL => '
+        UPDATE ticket
+        SET escalation_time = ?, escalation_response_time = ?, escalation_update_time = ?,
+            escalation_solution_time = ?, change_time = current_timestamp, change_by = ?
+        WHERE id = ?',
+    Bind => [
+        \$EscalationTime,
+        \$EscalationTime,
+        \$EscalationTime,
+        \$EscalationTime,
+        \'1',
+        \$TicketID1,
+    ],
 );
 
 $BackendObject->ValueSet(
@@ -1259,6 +1278,86 @@ my @Tests = (
         ExpectedReturnRemoteData => {
             Data => {
                 TicketID => $TicketID4,
+            },
+            Success => 1,
+        },
+        Operation => 'TicketSearch',
+    },
+    {
+        Name           => "Test EscalationTime " . $TestCounter++,
+        SuccessRequest => 1,
+        RequestData    => {
+            TicketEscalationTimeNewerMinutes => 120,
+        },
+        ExpectedReturnLocalData => {
+            Data => {
+                TicketID => [$TicketID1],
+            },
+            Success => 1
+        },
+        ExpectedReturnRemoteData => {
+            Data => {
+                TicketID => $TicketID1,
+            },
+            Success => 1,
+        },
+        Operation => 'TicketSearch',
+    },
+    {
+        Name           => "Test EscalationResponseTime " . $TestCounter++,
+        SuccessRequest => 1,
+        RequestData    => {
+            TicketEscalationResponseTimeNewerMinutes => 120,
+        },
+        ExpectedReturnLocalData => {
+            Data => {
+                TicketID => [$TicketID1],
+            },
+            Success => 1
+        },
+        ExpectedReturnRemoteData => {
+            Data => {
+                TicketID => $TicketID1,
+            },
+            Success => 1,
+        },
+        Operation => 'TicketSearch',
+    },
+    {
+        Name           => "Test EscalationUpdateTime " . $TestCounter++,
+        SuccessRequest => 1,
+        RequestData    => {
+            TicketEscalationUpdateTimeNewerMinutes => 120,
+        },
+        ExpectedReturnLocalData => {
+            Data => {
+                TicketID => [$TicketID1],
+            },
+            Success => 1
+        },
+        ExpectedReturnRemoteData => {
+            Data => {
+                TicketID => $TicketID1,
+            },
+            Success => 1,
+        },
+        Operation => 'TicketSearch',
+    },
+    {
+        Name           => "Test EscalationSolutionTime " . $TestCounter++,
+        SuccessRequest => 1,
+        RequestData    => {
+            TicketEscalationSolutionTimeNewerMinutes => 120,
+        },
+        ExpectedReturnLocalData => {
+            Data => {
+                TicketID => [$TicketID1],
+            },
+            Success => 1
+        },
+        ExpectedReturnRemoteData => {
+            Data => {
+                TicketID => $TicketID1,
             },
             Success => 1,
         },